Changes data types used in nre castor version for mappings

+++ Common/src/alma/scheduling/datamodel/obsproject/dao/XmlObsProjectDaoImpl.java
@@ -207,7 +207,7 @@ public class XmlObsProjectDaoImpl implements XmlObsProjectDao {
                     xmlSchedBlock.getSchedulingConstraints().getMaxAngularResolution(),
                     xmlSchedBlock.getSchedulingConstraints().getRepresentativeFrequency(),
                     targets.get(xmlSchedBlock.getSchedulingConstraints().getRepresentativeTargetIdRef()));
-            sc.setRepresentativeBand(xmlSchedBlock.getSchedulingConstraints().getRepresentativeBand());
+            sc.setRepresentativeBand((int)xmlSchedBlock.getSchedulingConstraints().getRepresentativeBand());
             schedBlock.setSchedulingConstraints(sc);
             Preconditions pc =
                 new Preconditions(xmlSchedBlock.getPreconditions().getMaxAllowedHA(),
@@ -245,7 +245,7 @@ public class XmlObsProjectDaoImpl implements XmlObsProjectDao {
                 sbc.setIndefiniteRepeat(sbControl.getIndefiniteRepeat());
                 sbc.setState(SchedBlockState.READY);
                 sbc.setSbMaximumTime(sbControl.getEstimatedExecutionTime());
-                sbc.setExecutionCount(sbControl.getNRepetitions());
+                sbc.setExecutionCount((int)sbControl.getNRepetitions());
                 schedBlock.setSchedBlockControl(sbc);
             }
         }
